% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_extratos.r
\name{e_cef_extratos}
\alias{e_cef_extratos}
\title{Consolida\u00e7\u00e3o dos dados dos extratos da CEF}
\usage{
e_cef_extratos(
  f_caminho.pasta.extratos_c = here("Relatórios - Documentos", "Relatorios - Extratos")
)
}
\arguments{
\item{f_caminho.pasta.extratos_c}{Caminho completo para a pasta que
cont\u00e9m os arquivos PDF dos extratos.}
}
\value{
Retorna um tibble com as seguintes colunas:
\itemize{
\item Data de lan\u00e7amento: Date.
\item Data de movimento: Date.
\item Documento: Character.
\item Hist\u00f3rico: Character.
\item Valor: Numeric.
\item Saldo: Numeric.
\item Conta_interno: Character.
\item Conta: Character.
\item Ag\u00eAncia: Character.
\item Produto: Character.
\item CNPJ: Character.
\item Cliente: Character.
\item Per\u00edodo_in\u00edcio: Date.
\item Per\u00edodo_fim: Date.
\item Data_consulta: POSIXct.
}
}
\description{
Consolida e processa dados de m\u00faltiplos extratos em PDF da CEF,
combinando-os em um \u00fanico data frame.
}
\details{
A fun\u00e7\u00e3o percorre a pasta especificada buscando arquivos PDF que
contenham os c\u00f3digos 2429, 2419 ou 2245, ignorando aqueles que contenham
a palavra \u0022fundo\u0022. Para cada arquivo encontrado, chama a fun\u00e7\u00e3o
\code{e_cef_extrato} para realizar a extra\u00e7\u00e3o dos dados e, posteriormente,
consolida os resultados em um \u00fanico tibble.
}
\examples{
\dontrun{
extratos <- e_cef_extratos(
  f_caminho.pasta.extratos_c = "caminho/para/a/pasta/dos/extratos"
)
print(extratos)
}

}
