% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_epr.R
\name{e_cef_epr}
\alias{e_cef_epr}
\title{Extra\u00e7\u00e3o dos dados do PDF de um relat\u00f3rio EPR da CEF}
\usage{
e_cef_epr(f_caminho.epr_c)
}
\arguments{
\item{f_caminho.epr_c}{Caminho completo para o arquivo PDF contendo o
relat\u00f3rio EPR da CEF.}
}
\value{
Retorna uma tibble com as seguintes colunas:
\itemize{
\item CONTRATO       : Character
\item NOME MUTUARIO  : Character
\item UNO            : Character
\item ORR            : Character
\item TO             : Integer
\item COD            : Character
\item DT. ASSIN      : Date
\item TIPO UND       : Character
\item GAR. AUT       : Integer
\item DT. INC. CTR   : Date
\item DT. INC. REG   : Date
\item VR RETIDO      : Numeric
\item VR AMORTIZ     : Numeric
\item AMO            : Character
}
}
\description{
Extrai e organiza dados de um relat\u00f3rio EPR da CEF em PDF.
}
\details{
Utiliza o pacote pdftools para ler o arquivo e manipular o texto,
identificando padr\u00f5es que auxiliam na extra\u00e7\u00e3o das informa\u00e7\u00f5es.
}
\examples{
\dontrun{
# Exemplo 1: Uso b\u00e1sico
epr <- e_cef_epr(f_caminho.epr_c = "caminho/para/o/relatorio_epr.pdf")
print(epr)

# Exemplo 2: Integrando com outras fun\u00e7\u00f5es de tratamento de dados
epr_filtrado <- e_cef_epr("caminho/para/o/relatorio_epr.pdf") \%>\%
  filter(`VR RETIDO` > 1000)
summary(epr_filtrado)
}

}
\references{
Consulte \code{\link{pdf_text}} para extra\u00e7\u00e3o de texto de arquivos PDF.
}
\seealso{
Consulte \code{\link{e_cef_eprs}}.
}
